feat : 보너스볼 2등 추가로 인한 view 추가 개발

+++ src/main/java/lotto/application/PlayLotto.java
@@ -48,7 +48,7 @@ public class PlayLotto {
         for (int index = 0; index < winningArray.length; index++) {
             totalMoney += calculate(index, winningArray[index]);
         }
-        return Math.round((((float) totalMoney / cash) * 100)) / 100.0;
+        return Math.round((((double) totalMoney / cash) * 100)) / 100.0;
     }
 
     private static int calculate(int winningIndex, int count) {
